+++ class/PDFStampData.class.php
@@ -1,0 +1,12 @@
+<?php
+
+class PDFStampData {
+
+    /* Date en ISO */
+
+    public $envoi_prefecture_date;
+    public $recu_prefecture_date;
+    public $affichage_date;
+    public $identifiant_unique;
+
+}