+++ CHANGELOG.md
@@ -69,6 +69,7 @@ d'un utilisateur. L'affichage de cette nouvelle page a été optimisé. #18 ~Adm
 - La tentative de création d'un groupe vide ne provoque plus d'erreur #198 ~Mailsec 
 - Ne pas pouvoir traiter les fichier d'un lot déjà traité #201 ~Actes 
 - Les jetons d'horodatage utilise le format ISO 8601 pour le message horodaté 
+- Remplacement complet du système de tampon des actes #202 ~Actes 
 
 ### Retraits